backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "tutorial project on backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>5.0,<6.0",
    "django-debug-toolbar==6.0.0",
    "djangorestframework==3.16.1",
    "django-cors-headers==4.9.0",
    "djangorestframework-simplejwt==5.5.1",
    "dj-rest-auth[with_social]==7.0.0",
    "drf-spectacular==0.28.0",
    "langchain-openai==0.3.33",
    "python-dotenv==1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.5.0",
    "pytest>=8.3.0",
]
